backend/app/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
flask-cors==6.0.2
gunicorn==23.0.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
mariadb==1.1.14
MarkupSafe==3.0.3
packaging==25.0
pycparser==2.23
pydantic_core==2.41.5
requests==2.32.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
